requirements.txt
opencv-python
numpy
mediapipe
pandas
pyyaml
sanic
matplotlib
pandas
jupyter
websockets==12.0